fix windows warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@57403 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: f476eb23d6f35d2399f4fa7f45a6a87d2375e935

+++ LHCb/L0/ProcessorKernel/src/Property.cpp
@@ -44,7 +44,7 @@ const std::vector<float> L0Muon::Property::getVectorOfFloat() const{
   char * endptr;
 
   for (int k=0;k<length;k++){
-    float f = strtod(str,&endptr);
+    float f = (float)strtod(str,&endptr);
     if (endptr==str) {
       break;
     }
